Move some database stuf to database storage
