+++ AirCasting/Utils/AveragingService.swift
@@ -43,7 +43,7 @@ enum AveragingWindow: Int {
 
 enum TimeThreshold: Int {
     // Two hours: 60 * 60 * 2 = 7200
-    case firstThreshold = 120 //TODO: change back
+    case firstThreshold = 7200
     // Nine hours: 60 * 60 * 9 = 32400
     case secondThreshold = 32400
 }
